feat: coerce column types in GetFields for UNION
