+++ go/vt/vtgate/planbuilder/operators/apply_join.go
@@ -175,8 +175,8 @@ func (aj *ApplyJoin) GetSelectExprs(ctx *plancontext.PlanningContext) sqlparser.
 	return transformColumnsToSelectExprs(ctx, aj)
 }
 
-func (aj *ApplyJoin) GetOrdering() []ops.OrderBy {
-	return aj.LHS.GetOrdering()
+func (aj *ApplyJoin) GetOrdering(ctx *plancontext.PlanningContext) []ops.OrderBy {
+	return aj.LHS.GetOrdering(ctx)
 }
 
 func joinColumnToAliasedExpr(c JoinColumn) *sqlparser.AliasedExpr {
