+++ go/vt/vtgate/planbuilder/operators/comments.go
@@ -76,6 +76,6 @@ func (l *LockAndComment) ShortDescription() string {
 	return strings.Join(s, " ")
 }
 
-func (l *LockAndComment) GetOrdering() []ops.OrderBy {
-	return l.Source.GetOrdering()
+func (l *LockAndComment) GetOrdering(ctx *plancontext.PlanningContext) []ops.OrderBy {
+	return l.Source.GetOrdering(ctx)
 }
