+++ go/vt/vtgate/planbuilder/operators/distinct.go
@@ -114,8 +114,8 @@ func (d *Distinct) ShortDescription() string {
 	return "Performance"
 }
 
-func (d *Distinct) GetOrdering() []ops.OrderBy {
-	return d.Source.GetOrdering()
+func (d *Distinct) GetOrdering(ctx *plancontext.PlanningContext) []ops.OrderBy {
+	return d.Source.GetOrdering(ctx)
 }
 
 func (d *Distinct) setTruncateColumnCount(offset int) {
