+++ go/vt/vtgate/planbuilder/operators/ops/op.go
@@ -53,7 +53,7 @@ type (
 
 		ShortDescription() string
 
-		GetOrdering() []OrderBy
+		GetOrdering(ctx *plancontext.PlanningContext) []OrderBy
 	}
 
 	// OrderBy contains the expression to used in order by and also if ordering is needed at VTGate level then what the weight_string function expression to be sent down for evaluation.
